+++ SmartLogs/__init__.py
@@ -1,5 +1,22 @@
 from .logger import SmartLogger
 from .debug import DebugLogger
+from .formatter import (
+    SmartFormatter,
+    JSONFormatter,
+    XMLFormatter,
+    YAMLFormatter,
+    TableFormatter,
+    TemplateFormatter
+)
 
-__version__ = "0.1.0"
-__all__ = ["SmartLogger", "DebugLogger"]
+__version__ = "0.1.1"
+__all__ = [
+    "SmartLogger",
+    "DebugLogger",
+    "SmartFormatter",
+    "JSONFormatter",
+    "XMLFormatter",
+    "YAMLFormatter",
+    "TableFormatter",
+    "TemplateFormatter"
+]